move tests to own project
