+++ server/package.json
@@ -7,8 +7,6 @@
     "start": "nodemon app.ts",
     "start:dev": "ts-node app.ts",
     "debug": "nodemon app.ts",
-    "test:unit": "vitest -c test/vitest.config.unit.ts",
-    "test:int": "./test/integration_test.sh",
     "insertData": "ts-node util/db-scripts/deleteData.ts && ts-node util/db-scripts/insertData.ts",
     "readData": "ts-node util/db-scripts/readData.ts",
     "deleteData": "ts-node util/db-scripts/deleteData.ts"
