+++ test/vitest.config.integration.ts
@@ -2,6 +2,6 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
-    include: ['./test/domain/*.test.ts', './test/routes/*.test.ts'],
+    include: ['./persistence/*.test.ts'],
   },
 });
