+++ test/vitest.config.unit.ts
@@ -2,6 +2,6 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
-    include: ['./test/persistence/*.test.ts'],
+    include: ['./domain/*.test.ts', './routes/*.test.ts'],
   },
 });
